Create DiviStock Inc.
A FinTech case study for a blockchain / cryptocurrency startup company that aims to create a platform that allow the tokenization of Assets. DiviStock is this new and upcoming success that offers peer to peer transactions with a  multitude of features and options to optimize the value of assets by means of the cryptocurrency techniques.